feat: save filter group to database

+++ dts-playlist-core/src/main/kotlin/com/gugu/dts/playlist/core/repository/MusicLibraryRepository.kt
@@ -9,7 +9,7 @@ interface MusicLibraryRepository {
     fun fetchLibraryByName(name: String): MusicLibrary?
     fun list(): List<MusicLibrary>
     fun deleteLibById(currentLibId: Long)
-    fun find(id: Long): IMusicLibrary?
+    fun find(id: Int): IMusicLibrary?
     fun resetPlayedTimes(libId: Long)
     fun updateSongPlayedTimes(id: Long, newValue: Int)
     fun updateSongPlayedTimesByOne(ids: LongArray)
